requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
mysqlclient==2.2.5
python-dotenv==1.0.1
sqlparse==0.5.1
tzdata==2024.2
djangorestframework-simplejwt
gunicorn
psycopg2-binary
dj-database-url
whitenoise
